Redux added
