--- client/app/check/page.tsx
@@ -1,9 +1,0 @@
-import React from 'react'
-
-const page = () => {
-  return (
-    <div>page</div>
-  )
-}
-
-export default page
