--- client/app/redux/slice/adminSlice.ts
@@ -1,1 +1,0 @@
-"use client"
