--- client/app/redux/slice/employeeSlice.ts
@@ -1,1 +1,0 @@
-"use client"
